fix(ui): A4 -- piety mostraba "4294967160" (mismatch %ld con s32b bajo L64)
prt_piety hacía sprintf("%9ld", p_ptr->grace) sin cast. grace es s32b (== int en
LP64/L64), pero %ld lee un long de 64 bits -> basura (p.ej. 4294967160 = un grace
negativo/pequeño malinterpretado). Cast a (long), como ya hacían exp y au en el
mismo sidebar. Ahora muestra la piedad real (incl. negativa).

Además: nota en docs/autoplay-port-notes.md del punto B (objetos lanzables/a
distancia) para el roadmap del bot, pendiente de hacer.

(cherry picked from commit 5ac778e2ca0650d2707792cb33e2cba85f908c2c)

+++ src/xtra1.c
@@ -146,7 +146,9 @@ static void prt_piety(void)
 
 	c_put_str(TERM_L_WHITE, "Pt ", ROW_PIETY, COL_PIETY);
 
-	sprintf(tmp, "%9ld", p_ptr->grace);
+	/* grace is s32b (== int under L64); %ld expects a 64-bit long, so without the
+	 * cast it reads garbage and the piety shows as e.g. "4294967160". */
+	sprintf(tmp, "%9ld", (long)p_ptr->grace);
 
 	if (p_ptr->praying)
 		c_put_str(TERM_L_GREEN, tmp, ROW_PIETY, COL_PIETY + 3);
